refactor(core): fold state-transition timers into overlay-effects
Two more identical-shape effects:

  current().state === 'justDragged' -> finishJustDragged after FEEDBACK_DURATION_MS
  current().state === 'justCopied'  -> finishJustCopied  after FEEDBACK_DURATION_MS

were sitting at top of init() body. Move them into createOverlayEffects
alongside the other auto-managed-by-isActivated timers and animations.
Drops FEEDBACK_DURATION_MS from index.tsx imports.

index.tsx: 2878 -> 2860 (-18).
